create a seperate utils for search list constants
